requirements.txt
hashids==1.2.0
hvac==0.5.0
pycryptodome==3.6.1
requests==2.18.4
